requirements.txt
folium==0.20.0
geopy==2.4.1
networkx==3.5
numpy==2.3.4
osmnx==2.0.6
PyYAML==6.0.2
PyYAML==6.0.3
tqdm==4.67.1
